# Fix wrong method lookup cache used when class is prepended