@headius 2017-05-03
Looks great, and thank you for such a detailed analysis and commit message!